Respond to PR comments
